fix: export types
